+++ Algorithms/BayesianOptimization/Vanilla_BO.py
@@ -91,7 +91,7 @@ class Vanilla_BO(AbstractBayesianOptimizer):
 
         self.__torch_config = {
             "device": torch.device("cuda:0" if torch.cuda.is_available() else "cpu"),
-            "dtype": torch.float,
+            "dtype": torch.double,
             "NUM_RESTARTS": 20,
             "RAW_SAMPLES": 4096,
             "OPTIMIZE_ACQF_OPTIONS": {"maxiter": 100, "method": "L-BFGS-B"},
